app/requirements.txt
Flask
mysql-connector
datetime
requests
asyncio
timedelta
